+++ src/org/minima/system/commands/help.java
@@ -24,10 +24,10 @@ import org.minima.system.commands.base.tokenvalidate;
 import org.minima.system.commands.base.trace;
 import org.minima.system.commands.base.tutorial;
 import org.minima.system.commands.base.vault;
+import org.minima.system.commands.maxima.maxcontacts;
+import org.minima.system.commands.maxima.maxima;
 import org.minima.system.commands.network.connect;
 import org.minima.system.commands.network.disconnect;
-import org.minima.system.commands.network.maxcontacts;
-import org.minima.system.commands.network.maxima;
 import org.minima.system.commands.network.message;
 import org.minima.system.commands.network.network;
 import org.minima.system.commands.network.rpc;
